Small fix for sign warning
git-svn-id: https://geany.svn.sourceforge.net/svnroot/geany/trunk@888 ea778897-0a13-0410-b9d1-a72fbfd435f5

+++ src/keyfile.c
@@ -524,7 +524,7 @@ gboolean configuration_open_files()
 		if (app->tab_order_ltr)
 		{
 			i++;
-			if (i >= session_files->len) break;
+			if (i >= (gint)session_files->len) break;
 		}
 		else
 		{
